+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -8,10 +8,10 @@ public class StatsServiceTest {
     @Test
     public void totalSalesYear() {
         StatsService service = new StatsService();
-        int[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int sum = 180;
-        int actualSum = service.salesAmount(monthsOfSales);
+        long sum = 180;
+        long actualSum = service.salesAmount(monthsOfSales);
 
         Assertions.assertEquals(sum, actualSum);
 
@@ -20,10 +20,10 @@ public class StatsServiceTest {
     @Test
     public void averageMonthlySalesForTheYear() {
         StatsService service = new StatsService();
-        int[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int averageAmount = 15;
-        int actualSum = service.averageSalesAmount(monthsOfSales);
+        long averageAmount = 15;
+        long actualSum = service.averageSalesAmount(monthsOfSales);
 
         Assertions.assertEquals(averageAmount, actualSum);
 
@@ -32,10 +32,10 @@ public class StatsServiceTest {
     @Test
     public void maxSalesForTheYear() {
         StatsService service = new StatsService();
-        int[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        long maxMonth = 8;
-        long actualSum = service.maxSales(monthsOfSales);
+        int maxMonth = 8;
+        int actualSum = service.maxSales(monthsOfSales);
 
         Assertions.assertEquals(maxMonth, actualSum);
 
@@ -44,10 +44,10 @@ public class StatsServiceTest {
     @Test
     public void minSalesForTheYear() {
         StatsService service = new StatsService();
-        int[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        long minMonth = 9;
-        long actualSum = service.minSales(monthsOfSales);
+        int minMonth = 9;
+        int actualSum = service.minSales(monthsOfSales);
 
         Assertions.assertEquals(minMonth, actualSum);
 
@@ -56,10 +56,10 @@ public class StatsServiceTest {
     @Test
     public void NumberOfMonthsOfSalesBelowAverage() {
         StatsService service = new StatsService();
-        int[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int numberOfMonths = 5;
-        int actualSum = service.numberOfMonthsOfSalesBelowAverage(monthsOfSales);
+        long numberOfMonths = 5;
+        long actualSum = service.numberOfMonthsOfSalesBelowAverage(monthsOfSales);
 
         Assertions.assertEquals(numberOfMonths, actualSum);
 
@@ -68,10 +68,10 @@ public class StatsServiceTest {
     @Test
     public void NumberOfMonthsOfSalesAboveAverage() {
         StatsService service = new StatsService();
-        int[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthsOfSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int numberOfMonths = 5;
-        int actualSum = service.numberOfMonthsOfSalesAboveAverage(monthsOfSales);
+        long numberOfMonths = 5;
+        long actualSum = service.numberOfMonthsOfSalesAboveAverage(monthsOfSales);
 
         Assertions.assertEquals(numberOfMonths, actualSum);
